backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy[asyncio]==2.0.29
asyncpg==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
